+++ utils/SaveFunction.py
@@ -16,14 +16,5 @@ def save(leaders_per_country: dict) -> None:
     with open(filename, "w") as output:
         output.write(now + "/n" + json.dumps(leaders_per_country))
 
-# Testing functions
-
-def test_save():
-    save({'Empty':'Dict'})
-    assert os.path.exists('./results/leaders.json')
-    with open('./results/leaders.json', 'r') as file:
-              text = file.read()
-              assert str(datetime.today().date()) == text[:10]
-
 if __name__ == "__main__":
-    pytest.main(['-v'])
+    pytest.main(['./test_SaveFunction.py', '-v'])
